+++ ideal-boot-project/ideal-boot-autoconfigure/src/main/java/cn/idealframework/boot/autoconfigure/event/broker/properties/rabbit/IdealBootEventRabbitProperties.java
@@ -30,6 +30,11 @@ public class IdealBootEventRabbitProperties {
   /** 消费队列前缀 */
   private String queuePrefix = "event.queue";
 
+  /**
+   * 是否开启本地模式
+   */
+  private boolean enableLocalModel = false;
+
   /** 消费者预取消息数 */
   private int prefetchCount = 0;